Implement fixed update

+++ engine/src/utils/types.zig
@@ -1,1 +1,1 @@
-pub const Deltatime = f32;
+pub const Deltatime = f64;
